MIM-171 按职责拆分排队 Turn 与审批投影
源码体积门禁在本 PR 中失败：SessionStoreTurns.swift 与
CodexAppServerEventProjection.swift 都从 1997 行越过 2000 行上限。

按职责拆出两个文件，纯代码搬移，不改行为：
- SessionStoreQueuedTurns.swift：排队 Turn 的调度、连接监控、
  发送结果归并与持久化对账。
- CodexAppServerApprovalProjection.swift：审批、用户输入与 MCP
  反向请求的挂起登记、校验与响应构造。
